+++ src/System/MQ/Protocol/Internal/Instances.hs
@@ -10,9 +10,11 @@ import           Data.ByteString                   as BS (ByteString)
 import           Data.Map.Strict                   (Map, fromList, member, (!))
 import           Data.MessagePack.Types.Class      (MessagePack (..))
 import           Data.MessagePack.Types.Object     (Object)
-import           System.MQ.Protocol.Internal.Types (Creator, Dictionary (..), MessageType (..),
+import           Data.Text                         (Text)
+import           System.MQ.Protocol.Internal.Types (Creator, Dictionary (..),
                                                     Encoding, Hash,
-                                                    Message (..), Spec,
+                                                    Message (..),
+                                                    MessageType (..), Spec,
                                                     Timestamp)
 
 infix .=
@@ -20,7 +22,7 @@ infix .=
 a .= b = (a, toObject b)
 
 infix .!
-(.!) :: (Monad m, MessagePack b) => Map ByteString Object -> ByteString -> m b
+(.!) :: (Monad m, MessagePack b) => Map Text Object -> Text -> m b
 dict .! key | key `member` dict = fromObject $ dict ! key
             | otherwise = error $ "System.MQ.Protocol.Internal.Instances: .! :: key " ++ show key ++ " is not an element of the dictionary."